Completely reorganize the sketchbook module.
