+++ src-tauri/src/sketchbook/model/_binary_op.rs
