+++ src-tauri/src/sketchbook/model/_monotonicity.rs
